// src/app/Features/display-board/models/sale.ts
export class Sale {
    SalesID?: number;
    Invoice?: string;
    SaleDT?: Date;
    SupID?: string;
    CategoryId?: string;
    Item_id?: string;
    ItemName?: string
    ItemSL?: number;
    OptionsID?: string;
    OptionsGroupName?: string;
    AddonsID?: string;
    CPU?: number;
    RPU?: number;
    RPU_Text?: string
    SQty?: number;
    Qty_Text?: string;
    VQty_Text?: string;
    rQty?: number;
    rAmt?: number;
    cInvoice?: string;
    ShopID?: string;
    PayType?: string
    TotalCost?: number;
    TotalAmt?: number;
    Discount?: number;
    DiscAmt?: number;
    DiscAmtPrd?: number;
    Sc?: number;
    ScAmt?: number;
    VAT?: number;
    NetAmt?: number;
    PaidAmt?: number;
    Salesman?: string;
    CounterID?: string;
    PrvCusID?: string;
    VATPrcnt?: number;
    DiscPrcnt?: number;
    Returned?: string;
    TblNo?: string;
    Persons?: number;
    Waiter?: string;
    KOT?: string;
    UserId?: string;
    ServiceCharge?: number;
    IsHappyHourPrice?: string;
    OrderNo?: string;
    Area?: string;
    Instruction?: string;
    InstructionIDs?: string;
    IsInvoicePrint?: string;
    Priority?: string;
    VoidReason?: string;
    ContinentalRPU?: number;
    Department?: number;
    isMemberPaid?: boolean;
}